Removendo produto e validando estoque antes

+++ src/main/java/br/com/stockflowservice/service/impl/ProdutoServiceImpl.java
@@ -108,6 +108,13 @@ public class ProdutoServiceImpl implements ProdutoService {
     @Override
     public void deletarProduto(ProdutoDTO produto) {
         Produto produtoEncontrado = this.buscarUmProduto(produto.getId());
+
+        Estoque estoque = estoqueService.buscarUmEstoque(produtoEncontrado);
+
+        if (estoque.getQuantidade() > 0) {
+            throw new StockFlowException("Esse produto possui estoque, é preciso dar baixa no estoque para excluir o produto!", HttpStatus.BAD_REQUEST);
+        }
+        estoqueService.deletarEstoque(estoque.getId());
         produtoRepository.delete(produtoEncontrado);
     }
 }
